requirements/requirements.txt
keyring~=18.0.1
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework-simplejwt
djangorestframework
python-dotenv